pyproject.toml
[project]
name = "plot-agent"
version = "0.1.0"
description = "Langgraph agent that reads data from a database and generates plots with captions and text analysis"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "baml-py==0.212.0",
    "click>=8.3.0",
    "fire>=0.7.1",
    "kaleido>=1.1.0",
    "langchain-community>=0.3.30",
    "langchain-experimental>=0.3.4",
    "langchain[google-genai]>=0.3.27",
    "langfuse>=3.6.1",
    "langgraph>=0.6.8",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "ipython>=9.6.0",
]
